Adds export to Edge adapter Extracts BuildProps class to lib file
