+++ packages/astro-sst/src/edge/adapter.ts
@@ -1,4 +1,5 @@
 import type { AstroAdapter, AstroIntegration } from "astro";
+import { BuildProps } from "../lib/buildProps";
 
 const PACKAGE_NAME = "astro-sst/edge";
 
@@ -14,8 +15,13 @@ export default function createIntegration(): AstroIntegration {
   return {
     name: PACKAGE_NAME,
     hooks: {
-      "astro:config:done": ({ setAdapter }) => {
+      "astro:config:done": ({ config, setAdapter }) => {
+        BuildProps.setAstroConfig(config);
         setAdapter(getAdapter());
+      },
+      "astro:build:done": async (buildResults) => {
+        BuildProps.setBuildResults(buildResults);
+        await BuildProps.exportBuildProps();
       },
     },
   };
